requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.5
CacheControl==0.12.6
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.2.1
django-extensions==2.2.8
django-filter==2.2.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
firebase-admin==5.0.3
future==0.18.2
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.4
google-cloud-storage==1.42.3
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.41.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.19.1
idna==3.2
msgpack==1.0.2
packaging==21.0
Pillow==9.3.0
proto-plus==1.19.2
protobuf==3.18.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyotp==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2019.3
PyYAML==5.3.1
reportlab==3.5.68
requests==2.26.0
rsa==4.7.2
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.26.6
webencodings==0.5.1
whitenoise==5.0.1
xhtml2pdf==0.2.5
